process, thread and module snapshots, if accessed when empty, are transparently populated by scanning the debuguees.
I was reluctant to code this feature because I wanted users to be more aware of the snapshots and how they worked, by forcing them to populate the snapshots manually. However it was causing confusion and problems in third party tools based on WinAppDbg, so I changed my mind, and now snapshots are a little bit "smarter" and easier to use, and you don't get weird bugs if you forget to initialize them properly.