Refactored crash_logger.py to use an XML configuration file instead of command line switches.
This is the first version. I'm still not quite convinced on the file format, but I wanted to commit a working version before making more changes. The plan is to also have a tool to generate the XML config files.

Also, on Win7 the process wasn't being killed properly when a crash occurred. It seems to be that pointing EIP to ExitProcess doesn't avoid the typical Windows crash dialog (at least when running 32 bits processes in a 64 bits machine). I'll have to look deeper into this, but for now a simple TerminateProcess() will do.