Refactored the buffer watch functionality to fix an API design flaw found by Felipe Manzano (@feliam).
The new API returns an identifier for each buffer watch defined. Then that identifier must be passed to the method that destroys the buffer watch. This allows you to safely define multiple overlapping buffer watches in all situations. (The old API would fail in some circumstances, especially when watching overlapping buffers of more than two pages in size).

Backwards compatibility was maintained, so scripts written for WinAppDbg 1.4 that didn't trigger this bug will still function properly. For those who do trigger the bug, now it tries to fail more gracefully than just raising an exception - but if your script was relying on that exception being thrown it will now fail. It's a small incompatibility but I think it's worth it for the vast majority of cases.