Some changes to the default behavior for breakpoint-related exceptions. This somewhat breaks compatibility with 1.4. The benefit is a simpler debugger with no event handler (or a very crude one) will get a more sensible default behavior when not in hostile mode.
For example, when running Notepad you may get sometimes a breakpoint or single step exception that should normally not happen outside a debugger. In previous versions of WinAppDbg we were forced to handle those cases, from now on the default is to simply ignore them.

No change was made to hostile mode since it's assumed the programmer MUST handle every case very carefully anyway. This is meant to ease the development of simple test scripts and the such.