Ok, hardware breakpoint hooks are not working for me now, but that could be unrelated to hooks in general. I'll tackle that a a separate problem.
I also changed the API (for now at least) to remove the dont_hook stuff, because doing it proper would require me to keep an internal list of hooks that have been set. If I do that, I'd need to also add methods to list the currently installed hooks. Sounds like a good idea, but I'm a bit overwhelmed with this refactoring right now, so I'll leave it for later - otherwise I'll never finish this.